voice_agent/requirements.txt
fastapi
uvicorn
python-dotenv
librosa
livekit
pydantic
pydub
pypinyin
scipy
sentence-transformers
setuptools
silero-vad
torch
transformers==4.48.0
TTS
unidecode
whisper
cn2an
eng_to_ipa
gradio
inflect
jieba
langid
requests
sounddevice
soundfile
numpy
openai
duckduckgo_search
beautifulsoup4